# specs/002-schedule-mode/contracts/api.yaml
openapi: 3.0.0
info:
  title: ViMax Web 调度模式与调度管理 API
  version: 1.0.0
  description: >
    调度模式与调度管理功能的后端 API 契约。
    (A) Web 端 API：backend 暴露给 frontend 的新增端点。
    (B) 调度服务对接契约：backend 调用外部 Vimax Scheduler 的端点映射（X-API-Key 鉴权）。
servers:
  - url: http://127.0.0.1:8000

# =====================================================================
# (A) Web 端 API（backend → frontend）
# =====================================================================
paths:
...
  /api/schedule/projects/{schedule_project_id}/tasks:
    get:
      operationId: listScheduleProjectTasks
      summary: 项目调度任务列表（透传调度服务）
      description: 返回调度服务侧某项目下的任务列表（含状态、失败原因、结果信息）。
      parameters:
        - name: schedule_project_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 任务列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleTask'
  /api/schedule/projects/{schedule_project_id}/priority:
    patch:
      operationId: setScheduleProjectPriority
      summary: 置顶/调整调度项目优先级（透传调度服务）
      description: priority=1 表示置顶（最高），priority=0 表示取消置顶（未设置）。
      parameters:
        - name: schedule_project_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [priority]
              properties:
                priority: { type: integer, minimum: 0, maximum: 99 }
      responses:
        '200':
          description: 优先级已更新
          content:
            application/json:
              schema:
                type: object
                properties:
                  project_id: { type: integer }
                  priority: { type: integer }
        '503':
          description: 调度服务不可用
  /api/schedule/projects/{schedule_project_id}/cancel:
    post:
      operationId: cancelScheduleProject
      summary: 取消调度项目（透传调度服务）
      description: 取消该项目下全部 QUEUED 任务（RUNNING 不受影响）。任务级取消经确认后同样调用本接口。
      parameters:
        - name: schedule_project_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 取消成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  cancelled: { type: boolean }
        '503':
          description: 调度服务不可用
...
  /api/schedule/results/sync:
    post:
      operationId: syncScheduleResults
      summary: 回传调度结果（保存到项目目录并生成 generation_result）
      description: >
        对列表中的成功任务逐个获取结果详情，调用 ComfyUI view API 拉取输出并保存到项目目录，
        落库 GenerationResult（prompt_id 去重）；失败任务标记失败原因，不生成成功结果。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                result_ids:
                  type: array
                  items: { type: string }
                  description: 待回传的 business_task_id 列表；为空表示全部未同步结果
      responses:
        '200':
          description: 回传结果汇总
          content:
            application/json:
              schema:
                type: object
                properties:
                  synced: { type: integer, description: 成功回传并生成 generation_result 数量 }
                  failed: { type: integer, description: 失败任务数量 }
                  skipped: { type: integer, description: 跳过（已存在/去重）数量 }

# =====================================================================
# (B) 调度服务对接契约（backend → Vimax Scheduler @ http://127.0.0.1:8001）
# =====================================================================
...
components:
  schemas:
...
    ScheduleTask:
      type: object
      properties:
        business_task_id: { type: string, description: 本地 prompt_id }
        project_id: { type: integer }
        status: { type: string, enum: [QUEUED, RUNNING, SUCCESS, FAILED] }
        failure_reason: { type: string, nullable: true }
        result_outputs: { type: object, nullable: true }
        result_view: { type: object, nullable: true, description: ComfyUI output view 信息 }
        has_sync: { type: boolean }
        retry_count: { type: integer }
...